Added type hint
TODO:
- Implement Event Sub
- Make DB async
- Use cache for autocomplete
